Adding preper test file location for workflow file
