+++ .github/workflows/main.yml
@@ -1,0 +1,36 @@
+name: Python CI
+
+on:
+  push:
+    branches: [ main, develop ]
+  pull_request:
+    branches: [ main, develop]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    env: # Defining environment variables using GitHub secrets
+      APP_TOKEN: ${{ secrets.APP_TOKEN }}
+      APP_TICKETS: ${{ secrets.APP_TICKETS }}
+      APP_MAX_TEMPERATURE: ${{ secrets.APP_MAX_TEMPERATURE }}
+      APP_MIN_TEMPERATURE: ${{ secrets.APP_MIN_TEMPERATURE }}
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v2
+
+    - name: Set up Python 3.8
+      uses: actions/setup-python@v2
+      with:
+        python-version: 3.8
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+
+    - name: Run tests
+      run: |
+        # Here you should run your tests, e.g.:
+        python -m unittest discover tests
